Feature/updates

+++ src/main/java/com/mf/minutefictionbackend/controllers/StoryController.java
@@ -5,10 +5,12 @@ import com.mf.minutefictionbackend.dtos.outputDtos.CommentOutputDto;
 import com.mf.minutefictionbackend.dtos.outputDtos.StoryOutputDto;
 import com.mf.minutefictionbackend.enums.StoryStatus;
 import com.mf.minutefictionbackend.services.CommentService;
+import com.mf.minutefictionbackend.services.SecurityService;
 import com.mf.minutefictionbackend.services.StoryService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -21,20 +23,23 @@ public class StoryController {
 
     private final StoryService storyService;
     private final CommentService commentService;
+    private final SecurityService securityService;
 
-    public StoryController(StoryService storyService, CommentService commentService) {
+    public StoryController(StoryService storyService, CommentService commentService, SecurityService securityService) {
         this.storyService = storyService;
         this.commentService = commentService;
+        this.securityService = securityService;
     }
 
 
     // MANAGE SUBMITTING/SUBMITTED STORIES
 
-    @PreAuthorize("hasAnyAuthority('AUTHOR', 'EDITOR')")
     @PostMapping("/submit/{themeId}")
-    public ResponseEntity<StoryOutputDto> submitStory(@Valid @PathVariable Long themeId, @RequestParam String username, @RequestBody StoryInputDto storyInputDto) {
-        StoryOutputDto storyDto = storyService.submitStory(storyInputDto, themeId, username);
+    public ResponseEntity<StoryOutputDto> submitStory(@PathVariable Long themeId, @Valid @RequestBody StoryInputDto storyInputDto) {
+        securityService.checkIsAuthor();
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
 
+        StoryOutputDto storyDto = storyService.submitStory(storyInputDto, themeId, username);
         URI uri = URI.create(ServletUriComponentsBuilder
                 .fromCurrentRequest()
                 .path("/" + storyDto.getId()).toUriString());
@@ -42,80 +47,87 @@ public class StoryController {
         return ResponseEntity.created(uri).body(storyDto);
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
-    @PutMapping("/editor/{storyId}/update")
-    public ResponseEntity<StoryOutputDto> updateStory(@Valid @PathVariable("storyId") Long storyId, @RequestBody StoryInputDto updatedStory) {
+    @PatchMapping("/editor/{storyId}/update")
+    public ResponseEntity<StoryOutputDto> updateStory(@PathVariable("storyId") Long storyId, @Valid @RequestBody StoryInputDto updatedStory) {
+        securityService.checkIsEditor();
         StoryOutputDto updatedStoryDto = storyService.updateStory(storyId, updatedStory);
         return ResponseEntity.ok().body(updatedStoryDto);
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
-    @GetMapping("/editor/submitted")
-    public ResponseEntity<List<StoryOutputDto>> getAllSubmittedStories() {
-        List<StoryOutputDto> stories = storyService.getStoriesByStatus(StoryStatus.SUBMITTED);
-        return ResponseEntity.ok(stories);
-    }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
-    @GetMapping("/editor/stories")
-    public ResponseEntity<List<StoryOutputDto>> getStoriesByStatusAndThemeId(@RequestParam StoryStatus status, @RequestParam Long themeId) {
+    @GetMapping("/editor/overview")
+    public ResponseEntity<List<StoryOutputDto>> getStoriesByStatusAndThemeId(@RequestParam(required = false) StoryStatus status, @RequestParam(required = false) Long themeId) {
+        securityService.checkIsEditor();
         List<StoryOutputDto> stories = storyService.getStoriesByStatusAndThemeId(status, themeId);
         return ResponseEntity.ok(stories);
     }
 
-    @PreAuthorize("hasAuthority('EDITOR') or @securityService.isAuthor(#storyId)")
     @GetMapping("/{storyId}")
     public ResponseEntity<StoryOutputDto> getStoryById(@PathVariable("storyId") Long storyId) {
+        securityService.checkIsEditorOrAuthor(storyId);
+
         StoryOutputDto storyDto = storyService.getStoryById(storyId);
         return ResponseEntity.ok(storyDto);
     }
 
-    @PreAuthorize("hasAuthority('EDITOR') or @securityService.isAuthor(#storyId)")
     @DeleteMapping("/{storyId}")
     public ResponseEntity<Void> deleteStory(@PathVariable("storyId") Long storyId) {
+        securityService.checkIsEditorOrAuthor(storyId);
+
         storyService.deleteStoryById(storyId);
         return ResponseEntity.noContent().build();
     }
 
 
-
     // MANAGE ACCEPTING AND DECLINING STORIES
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @PatchMapping("/editor/{storyId}/accept")
     public ResponseEntity<Void> acceptStory(@PathVariable("storyId") Long storyId) {
+        securityService.checkIsEditor();
+
         storyService.acceptStory(storyId);
         return ResponseEntity.noContent().build();
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @PatchMapping("/editor/{storyId}/decline")
     public ResponseEntity<Void> declineStory(@PathVariable("storyId") Long storyId) {
+        securityService.checkIsEditor();
+
         storyService.declineStory(storyId);
         return ResponseEntity.noContent().build();
     }
-    
+
+    @PatchMapping("/editor/{storyId}/status")
+    public ResponseEntity<Void> updateStoryStatus(@PathVariable("storyId") Long storyId, @RequestParam StoryStatus status) {
+        securityService.checkIsEditor();
+
+        storyService.updateStoryStatus(storyId, status);
+        return ResponseEntity.noContent().build();
+    }
+
 
     // MANAGE PUBLISHING/PUBLISHED STORIES
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @PatchMapping("/editor/{storyId}/publish")
     public ResponseEntity<Void> publishStory(@PathVariable Long storyId) {
+        securityService.checkIsEditor();
+
         storyService.publishStory(storyId);
         return ResponseEntity.noContent().build();
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @PatchMapping("/editor/themes/{themeId}/publish")
     public ResponseEntity<Void> publishAllAcceptedStoriesByTheme(@PathVariable("themeId") Long themeId) {
+        securityService.checkIsEditor();
+
         storyService.publishAllAcceptedStoriesByTheme(themeId);
         return ResponseEntity.noContent().build();
     }
 
 
     @GetMapping("/published")
-    public ResponseEntity<List<StoryOutputDto>> getAllPublishedStories() {
-        List<StoryOutputDto> stories = storyService.getAllPublishedStoriesByDateDesc(StoryStatus.PUBLISHED);
+    public ResponseEntity<List<StoryOutputDto>> getAllPublishedStories(@RequestParam(required = false, defaultValue = "10") int limit, @RequestParam(required = false, defaultValue ="0") int offset) {
+        List<StoryOutputDto> stories = storyService.getAllPublishedStoriesByDateDescWithPagination(StoryStatus.PUBLISHED, limit, offset);
         return ResponseEntity.ok(stories);
     }
 
@@ -125,23 +137,10 @@ public class StoryController {
         return ResponseEntity.ok(storyDto);
     }
 
-
-
-    // MANAGE STORIES BY THEME
-
-
-
     @GetMapping("/published/themes/{themeName}")
     public ResponseEntity<List<StoryOutputDto>> getPublishedStoriesByThemeName(@PathVariable("themeName") String themeName) {
         List<StoryOutputDto> stories = storyService.getStoriesByStatusAndThemeName(StoryStatus.PUBLISHED, themeName);
         return ResponseEntity.ok(stories);
-    }
-
-    @PreAuthorize("hasAuthority('EDITOR')")
-    @GetMapping("/editor/{themeId}/stories")
-    public ResponseEntity<List<StoryOutputDto>> getStoriesByThemeOfAllStatus(@PathVariable("themeId") Long themeId) {
-        List<StoryOutputDto> relatedStories = storyService.getStoriesByTheme(themeId);
-        return ResponseEntity.ok(relatedStories);
     }
